+++ bzt/jmx.py
@@ -879,7 +879,7 @@ class JMX(object):
         return element
 
     @staticmethod
-    def _get_csv_config(path, delimiter, is_quoted, on_eof):
+    def _get_csv_config(path, delimiter, is_quoted, loop):
         """
 
         :type path: str
@@ -890,20 +890,11 @@ class JMX(object):
         """
         element = etree.Element("CSVDataSet", guiclass="TestBeanGUI",
                                 testclass="CSVDataSet", testname="CSV %s" % os.path.basename(path))
-        if on_eof == "loop":
-            recycle = True
-            stop = False
-        elif on_eof == "stop":
-            recycle = False
-            stop = True
-        else:
-            recycle = False
-            stop = False
         element.append(JMX._string_prop("filename", path))
         element.append(JMX._string_prop("delimiter", delimiter))
         element.append(JMX._bool_prop("quotedData", is_quoted))
-        element.append(JMX._bool_prop("recycle", recycle))
-        element.append(JMX._bool_prop("stopThread", stop))
+        element.append(JMX._bool_prop("recycle", loop))
+        element.append(JMX._bool_prop("stopThread", not loop))
 
         return element